requirements.txt
# Зависимости для программы переименования медиафайлов
# Установка: pip install -r requirements.txt

# Обязательные зависимости
Pillow>=10.0.0

# Опциональные зависимости
# ffmpeg - для извлечения метаданных из видео
# Установка Windows: скачайте с https://ffmpeg.org/download.html
# После установки ffmpeg должен быть доступен в PATH

# Для создания EXE файла (опционально)
# pyinstaller>=5.0
